# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LINKY API — Pay-to-Speak on Base
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experts:
    get:
      operationId: listExperts
      tags: [experts]
      summary: List recent experts
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expert"
    post:
      operationId: createExpert
      tags: [experts]
      summary: Create or claim an expert profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expert"
        "409":
          description: Handle or wallet already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /experts/{handle}:
    get:
      operationId: getExpert
      tags: [experts]
      summary: Public expert profile
      parameters:
        - in: path
          name: handle
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: linky-session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MessagingPlatform:
      type: string
      enum: [whatsapp, telegram]
    MessagingChannel:
      type: object
      description: Public-safe view of a seller's messaging channel — no raw handle.
      required: [enabled, platform, priceUsdc]
      properties:
        enabled: { type: boolean }
        platform:
          $ref: "#/components/schemas/MessagingPlatform"
        priceUsdc:
          type: string
          description: USDC amount as a decimal string
    CallsChannel:
      type: object
      description: Public-safe view of a seller's calls channel — no raw booking URL.
      required: [enabled, priceUsdc]
      properties:
        enabled: { type: boolean }
        priceUsdc:
          type: string
          description: USDC amount as a decimal string
    MessagingChannelInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        platform:
          $ref: "#/components/schemas/MessagingPlatform"
        handle:
          type: string
          description: WhatsApp phone number (E.164) or Telegram username
        priceUsdc: { type: string }
    CallsChannelInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        bookingUrl: { type: string }
        priceUsdc: { type: string }
    Expert:
      type: object
      description: Public expert profile — never includes raw contact details for paid channels.
      required:
        - id
        - handle
        - wallet
        - name
        - about
        - createdAt
      properties:
        id: { type: string }
        handle: { type: string }
        wallet: { type: string }
        name: { type: string }
        about:
          type: string
          description: Profession / one-liner shown publicly
        linkedinUrl:
          type: ["string", "null"]
        xUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        messaging:
          oneOf:
            - $ref: "#/components/schemas/MessagingChannel"
            - type: "null"
        calls:
          oneOf:
            - $ref: "#/components/schemas/CallsChannel"
            - type: "null"
        createdAt: { type: string }
    ExpertInput:
      type: object
      required:
        - handle
        - wallet
        - signature
        - timestamp
        - name
        - about
      properties:
        handle:
          type: string
          minLength: 2
          maxLength: 32
          pattern: "^[a-z0-9-]+$"
        wallet: { type: string }
        signature: { type: string }
        timestamp: { type: integer }
        name: { type: string, minLength: 1 }
        about: { type: string, minLength: 1 }
        linkedinUrl:
          type: ["string", "null"]
        xUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        messaging:
          oneOf:
            - $ref: "#/components/schemas/MessagingChannelInput"
            - type: "null"
        calls:
          oneOf:
            - $ref: "#/components/schemas/CallsChannelInput"
            - type: "null"
...